refactor!: move all_capitalisations_of to application crate
