+++ app/src/app.rs
@@ -24,7 +24,7 @@ use egui_extras::{Size, StripBuilder};
 use preview::{DesignPreview, MAX_ZOOM_LEVEL, MIN_ZOOM_LEVEL};
 
 use seance::{
-    all_capitalisations_of, cut_file, default_passes,
+    cut_file, default_passes,
     svg::{parse_svg, SVG_UNITS_PER_MM},
     DesignFile, PrintDevice, SendToDeviceError, ToolPass, BED_HEIGHT_MM, BED_WIDTH_MM,
 };
@@ -1895,5 +1895,62 @@ fn focus_changing(
                 }
             }
         });
+    }
+}
+
+/// Gets all of the possible capitalisations of a string.
+/// We need this because the library we use for showig file dialogs is not very clever,
+/// it does not match against file extensions case-insensitively. Therefore, we provide
+/// the file dialog library with all of the possible capitalisations of the file extensions
+/// we care about, just in case folks have bizarrely capitalised file extensions.
+///
+/// # Arguments
+/// * `input`: The string to generate all the capitalisations of.
+///
+/// # Returns
+/// An array of strings containing all of the possible capitalisations of the input string.
+pub fn all_capitalisations_of(input: &str) -> Vec<String> {
+    let mut result = vec![];
+
+    let bitmask = ((2_u32.pow(input.len() as u32)) - 1) as usize;
+
+    for mask in 0..=bitmask {
+        let mut new_str = String::new();
+        for i in 0..input.len() {
+            if mask & (1 << i) > 0 {
+                new_str += &input
+                    .chars()
+                    .nth(i)
+                    .expect(&format!("Could not get character {i}"))
+                    .to_uppercase()
+                    .to_string();
+            } else {
+                new_str += &input
+                    .chars()
+                    .nth(i)
+                    .expect(&format!("Could not get character {i}"))
+                    .to_lowercase()
+                    .to_string();
+            }
+        }
+        result.push(new_str);
+    }
+
+    result
+}
+
+#[cfg(test)]
+mod test {
+    use super::all_capitalisations_of;
+
+    #[test]
+    fn capitalisations() {
+        let mut result = all_capitalisations_of("svg");
+        result.sort();
+        assert_eq!(result.len(), 8);
+        assert_eq!(
+            result,
+            vec!["SVG", "SVg", "SvG", "Svg", "sVG", "sVg", "svG", "svg"]
+        )
     }
 }
